Move tests to the `tests` directory to allow vendoring into m-c
